+++ Steps/Azure/ServiceBus/src/ReadMessageFromAzureOptions.cs
@@ -1,7 +1,7 @@
 ﻿// <copyright file="ReadMessageFromAzureOptions.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2019-2020 Chris Trout
+// Copyright(c) 2019-2022 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -23,6 +23,7 @@
 // </copyright>
 namespace MyTrout.Pipelines.Steps.Azure.ServiceBus
 {
+    using global::Azure.Messaging.ServiceBus;
     using System;
     using System.Collections.Generic;
 
@@ -44,11 +45,6 @@ namespace MyTrout.Pipelines.Steps.Azure.ServiceBus
         /// Gets or sets the connection string for Azure Service Bus.
         /// </summary>
         public string AzureServiceBusConnectionString { get; set; } = string.Empty;
-
-        /// <summary>
-        /// Gets or sets UserProperty names that will be added to the Message from <see cref="MyTrout.Pipelines.Core.PipelineContext" />, if they are available.
-        /// </summary>
-        public IEnumerable<string> ApplicationProperties { get; set; } = new List<string>();
 
         /// <summary>
         /// Gets or sets the number of messages to download for each request to Azure Service Bus.
@@ -76,6 +72,11 @@ namespace MyTrout.Pipelines.Steps.Azure.ServiceBus
                 return new ServiceBusReadEntity(this.EntityPath);
             }
         }
+
+        /// <summary>
+        /// Gets or sets the prefix for keys of <see cref="ServiceBusReceivedMessage.ApplicationProperties"/> when copied to <see cref="IPipelineContext.Items"/>.
+        /// </summary>
+        public string MessageContextItemsPrefix { get; set; } = "MSG_";
 
         /// <summary>
         /// Gets or sets the time this step will wait for a new message until it times out.