+++ Steps/Azure/ServiceBus/test/ReadMessageFromStepOptionsTests.cs
@@ -41,28 +41,30 @@ namespace MyTrout.Pipelines.Steps.Azure.ServiceBus.Tests
             int batchSize = 100;
             int deliveryAttemptsBeforeDeadLetter = 2;
             string entityPath = "topic/subscription";
+            string messagePrefix = "CHANGED_";
             TimeSpan timeToWaitBetweenMessageChecks = new TimeSpan(0, 0, 1);
             TimeSpan timeToWaitForNewMessage = new TimeSpan(0, 1, 0);
+
 
             // act
             var result = new ReadMessageFromAzureOptions
             {
-                ApplicationProperties = applicationProperties,
                 AzureServiceBusConnectionString = azureServiceBusConnectionString,
                 BatchSize = batchSize,
                 DeliveryAttemptsBeforeDeadLetter = deliveryAttemptsBeforeDeadLetter,
                 EntityPath = entityPath,
+                MessageContextItemsPrefix = messagePrefix,
                 TimeToWaitBetweenMessageChecks = timeToWaitBetweenMessageChecks,
                 TimeToWaitForNewMessage = timeToWaitForNewMessage
             };
 
             // assert
             Assert.IsNotNull(result);
-            Assert.AreEqual(applicationProperties, result.ApplicationProperties);
             Assert.AreEqual(azureServiceBusConnectionString, result.AzureServiceBusConnectionString);
             Assert.AreEqual(batchSize, result.BatchSize);
             Assert.AreEqual(deliveryAttemptsBeforeDeadLetter, result.DeliveryAttemptsBeforeDeadLetter);
             Assert.AreEqual(entityPath, result.EntityPath);
+            Assert.AreEqual(messagePrefix, result.MessageContextItemsPrefix);
             Assert.AreEqual(timeToWaitBetweenMessageChecks, result.TimeToWaitBetweenMessageChecks);
             Assert.AreEqual(timeToWaitForNewMessage, result.TimeToWaitForNewMessage);